[VecGeom-87] Attempt to fix FillUncontainedPoints() and some debugging printouts
Details: Navigation benchmark is failing to find uncontained points when the reference
volume is translated with respect to its own mother volume.

I tried to fix this problem in FillUncontainedPoints(), by replacing
Contains() with UnplacedContains(), but I noticed a similar use case
in FillContainedPoints():

      bool inside = (placed) ? (*v)->Contains(points[i])
                             : (*v)->UnplacedContains(points[i]);

and I am trying to understand whether the same code applies in FillUncontainedPoints().

This is Work In Progress...