[VecGeom-35] Reducing nrep parameter to speedup unit tests
BTW, the performance is pretty linear on the number of repetitions, so 4 repetitions
is good enough to get a good performance estimate.  If increasing processing time is
needed because total time is too small, increasing the number of points provides better
averaging over shape volumes, so it is better than just increasing the number of
repetitions.