Fix VecGeom UMESIMD backend types and functions
The type traits work to let VecCore define the correct backend functions.
However, it must be noted that the wrapped vecgeom::UmeSimdMask is not
produced when the user makes a condition, like a < b. Instead, in this
case, a UME::SIMD::SIMDVecMask<N> is created. Hence that's the type that
should be in the type traits. For the same reason, backend interface
functions must be overloaded for types that only exist in VecGeom, e.g.
vecgeom::UmeSimdPrecisionVector, etc, because the backend in VecCore
only defines MaskedAssign and Blend for the pure UME::SIMD::SIMDVec_*
types. Note that these are overloads, not specializations, so there
should not be a "template <>" in front of these functions. Finally,
any overload should use vecgeom::kVectorSize, not a hard-coded value,
or things will break when the architecture changes. That's probably
why the scalar plugin was failing.