Use constructors to initialize vector types
UMESIMD has explicit constructors for vector types, which means that
if VecGeom is to work with UMESIMD, one cannot do

Real_v x = 1.0;

as 1.0 is a scalar value, and there is no implicit conversion allowed by
UMESIMD. Therefore, all conversions from scalars to vectors must be
explicit. That is, 1.0 has to be converted to Real_v(1.0). Vc does not
have this limitation, and neither do the other backends, so extra care
must be taken to avoid this kind of problem in the future.