Fix/prevent compiler warning
This commit is preventing a compiler warning about potential
overflow of unsigned long long numbers.

It is safe to be ignored here since the overflow cannot physically appear
on any system with less than petabytes of RAM.
The warning is only taken out from very particular helper files.

For completeness, warnings of the following type are fixed:

- SpecializedPlacedVolImplHelper.h:261:19: warning: iteration 2305843009213693951
  .... invokes undefined behavior [-Waggressive-loop-optimizations]
  ....   vecCore::Store(result, output + i);

These warnings occured for the Scalar backend