Remove cuda_rdc_find_package_config
This function is a simple wrapper, and it should be left up to clients
to call `find_package_handle_standard_args` in CONFIG_MODE after `find_package`.

Use of this command within a -Config.cmake file can lead to unexpected
behaviour in combination with check_required_components. Revisit at a
later date.